Use path alias and move files into src directory
