+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "objective-manager-app",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "src/App.tsx",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -15,19 +15,20 @@
     "@react-navigation/native-stack": "^6.7.0",
     "expo": "^46.0.0",
     "expo-status-bar": "~1.4.0",
+    "expo-updates": "~0.14.4",
     "react": "18.0.0",
     "react-dom": "18.0.0",
     "react-native": "0.69.4",
     "react-native-safe-area-context": "4.3.1",
     "react-native-screens": "~3.15.0",
-    "react-native-web": "~0.18.7",
-    "expo-updates": "~0.14.4"
+    "react-native-web": "~0.18.7"
   },
   "devDependencies": {
     "@babel/core": "^7.18.6",
     "@react-native-community/eslint-config": "^3.1.0",
     "@types/react": "~18.0.0",
     "@types/react-native": "~0.69.1",
+    "babel-plugin-module-resolver": "^4.1.0",
     "eslint": "^8.21.0",
     "prettier": "^2.7.1",
     "typescript": "^4.6.3"
